replaced Makefile with doit and dodo.py files
