+++ python/dodo.py
@@ -1,0 +1,8 @@
+from open_discourse.steps.contributions._dodo import task_07_contributions
+from open_discourse.steps.data._dodo import task_01_download
+from open_discourse.steps.database._dodo import task_08_upload
+from open_discourse.steps.electoral_term_20._dodo import task_06_electoral_term_20
+from open_discourse.steps.factions._dodo import task_03_factions
+from open_discourse.steps.politicians._dodo import task_04_politicians
+from open_discourse.steps.preprocessing._dodo import task_02_preprocessing
+from open_discourse.steps.speech_content._dodo import task_05_speech_content
