+++ python/pyproject.toml
@@ -17,6 +17,7 @@ dependencies = [
     "pyparsing>=3.2.1",
     "pydantic>=2.10.6",
     "psycopg2-binary>=2.9.10",
+    "doit>=0.36.0",
 ]
