+++ python/src/open_discourse/steps/data/download_MdB_data.py
@@ -5,18 +5,20 @@ import requests
 
 from open_discourse.definitions import path
 
-mp_base_data_link = "https://www.bundestag.de/resource/blob/472878/7d4d417dbb7f7bd44508b3dc5de08ae2/MdB-Stammdaten-data.zip"
 
-print("Download & unzip 'MP_BASE_DATA'...", end="", flush=True)
+def main(task):
+    mp_base_data_link = (
+        "https://www.bundestag.de/resource/blob/472878/7d4d417dbb7f7bd44508b3dc5de08ae2/MdB-Stammdaten-data.zip"
+    )
 
-r = requests.get(mp_base_data_link)
+    print("Download & unzip 'MP_BASE_DATA'...", end="", flush=True)
 
-with zipfile.ZipFile(io.BytesIO(r.content)) as z:
-    mp_base_data_path = path.DATA_RAW / "MP_BASE_DATA"
-    mp_base_data_path.mkdir(parents=True, exist_ok=True)
+    r = requests.get(mp_base_data_link)
 
-    z.extractall(mp_base_data_path)
+    with zipfile.ZipFile(io.BytesIO(r.content)) as z:
+        mp_base_data_path = path.MP_BASE_DATA
+        mp_base_data_path.mkdir(parents=True, exist_ok=True)
 
-assert (mp_base_data_path / "MDB_STAMMDATEN.XML").exists()
+        z.extractall(mp_base_data_path)
 
-print("Script 01_02 done.")
+    return True
