+++ python/src/open_discourse/steps/factions/create.py
@@ -9,25 +9,27 @@ POLITICIANS_STAGE_01 = path.POLITICIANS_STAGE_01
 FACTIONS_STAGE_01 = path.FACTIONS_STAGE_01
 FACTIONS_STAGE_01.mkdir(parents=True, exist_ok=True)
 
-# read data.
-mps = pd.read_pickle(POLITICIANS_STAGE_01 / "mps.pkl")
 
-factions = mps.loc[mps["institution_type"] == "Fraktion/Gruppe", "institution_name"]
+def main(task):
+    # read data.
+    mps = pd.read_pickle(POLITICIANS_STAGE_01 / "mps.pkl")
 
-unique_factions = np.unique(factions)
-unique_factions = np.append(
-    unique_factions,
-    [
-        "Südschleswigscher Wählerverband",
-        "Gast",
-        "Gruppe Nationale Rechte",
-        "Deutsche Soziale Union",
-    ],
-)
+    factions = mps.loc[mps["institution_type"] == "Fraktion/Gruppe", "institution_name"]
 
-unique_factions = pd.DataFrame(unique_factions, columns=["faction_name"])
+    unique_factions = np.unique(factions)
+    unique_factions = np.append(
+        unique_factions,
+        [
+            "Südschleswigscher Wählerverband",
+            "Gast",
+            "Gruppe Nationale Rechte",
+            "Deutsche Soziale Union",
+        ],
+    )
 
-save_path_factions = FACTIONS_STAGE_01 / "factions.pkl"
-unique_factions.to_pickle(save_path_factions)
+    unique_factions = pd.DataFrame(unique_factions, columns=["faction_name"])
 
-print("Script 03_01 done.")
+    save_path_factions = FACTIONS_STAGE_01 / "factions.pkl"
+    unique_factions.to_pickle(save_path_factions)
+
+    return True
