+++ python/src/open_discourse/steps/preprocessing/split_xml_electoral_term_20.py
@@ -11,34 +11,38 @@ INPUT_PATH = path.DATA_CACHE / "electoral_term_pp20" / "stage_01"
 # output directory
 OUTPUT_PATH = path.DATA_CACHE / "electoral_term_pp20" / "stage_02"
 
-for xml_file_path in tqdm(sorted(INPUT_PATH.glob("*.xml")), desc="Parsing term 20..."):
-    # read data
-    tree = et.parse(xml_file_path)
-    root = tree.getroot()
 
-    toc = et.ElementTree(root.find("vorspann"))
-    session_content = et.ElementTree(root.find("sitzungsverlauf"))
-    appendix = et.ElementTree(root.find("anlagen"))
-    meta_data = et.ElementTree(root.find("rednerliste"))
+def main(task):
+    for xml_file_path in tqdm(
+        sorted(INPUT_PATH.glob("*.xml")), desc="Parsing term 20..."
+    ):
+        # read data
+        tree = et.parse(xml_file_path)
+        root = tree.getroot()
 
-    save_path = OUTPUT_PATH / regex.search(r"\d+", xml_file_path.stem).group()
-    save_path.mkdir(parents=True, exist_ok=True)
-    # save to xmls
-    toc.write(save_path / "toc.xml", encoding="UTF-8", xml_declaration=True)
-    session_content.write(
-        save_path / "session_content.xml",
-        encoding="UTF-8",
-        xml_declaration=True,
-    )
-    appendix.write(
-        save_path / "appendix.xml",
-        encoding="UTF-8",
-        xml_declaration=True,
-    )
-    meta_data.write(
-        save_path / "meta_data.xml",
-        encoding="UTF-8",
-        xml_declaration=True,
-    )
+        toc = et.ElementTree(root.find("vorspann"))
+        session_content = et.ElementTree(root.find("sitzungsverlauf"))
+        appendix = et.ElementTree(root.find("anlagen"))
+        meta_data = et.ElementTree(root.find("rednerliste"))
 
-print("Script 02_03 done.")
+        save_path = OUTPUT_PATH / regex.search(r"\d+", xml_file_path.stem).group()
+        save_path.mkdir(parents=True, exist_ok=True)
+        # save to xmls
+        toc.write(save_path / "toc.xml", encoding="UTF-8", xml_declaration=True)
+        session_content.write(
+            save_path / "session_content.xml",
+            encoding="UTF-8",
+            xml_declaration=True,
+        )
+        appendix.write(
+            save_path / "appendix.xml",
+            encoding="UTF-8",
+            xml_declaration=True,
+        )
+        meta_data.write(
+            save_path / "meta_data.xml",
+            encoding="UTF-8",
+            xml_declaration=True,
+        )
+
+    return True
